Update method of creating a holding div that fits the viewport
The issue with previously using absolute positioning for this is that the div was not working within the flow of the document. So to fix this I used:

- relative positioning with 100% width and 100vh and
- jquery to provide the height of the div once the document had loaded for browsers that do not support the vh units

This means that the div will always get the correct height and on MOST browsers there will be no flash of a zero height div.